chore: move junit processing into its own module
We will have more "mergify ci" command in the future lets prepare that.
